+++ helper/src/main/java/com/vsg/helper/common/util/viewModel/IViewModelAllTextSearchRelation.kt
@@ -3,5 +3,5 @@ package com.vsg.helper.common.util.viewModel
 import androidx.lifecycle.LiveData
 
 interface IViewModelAllTextSearchRelation {
-    fun viewModelGetAllTextSearch(idRelation: Long): LiveData<List<String>>
+    fun viewModelGetAllTextSearch(idRelation: Int): LiveData<List<String>>
 }
