+++ helper/src/main/java/com/vsg/helper/common/util/viewModel/IViewModelNextCode.kt
@@ -1,5 +1,5 @@
 package com.vsg.helper.common.util.viewModel
 
 interface IViewModelNextCode {
-    fun viewModelNextAutoCode(idRelation: Long): Long
+    fun viewModelNextAutoCode(idRelation: Int): Int
 }
