+++ helper/src/main/java/com/vsg/helper/common/util/viewModel/IViewModelView.kt
@@ -3,5 +3,5 @@ package com.vsg.helper.common.util.viewModel
 import com.vsg.helper.common.model.IEntity
 
 interface IViewModelView<TEntity> where TEntity : IEntity {
-    fun viewModelView(id: Long): TEntity?
+    fun viewModelView(id: Int): TEntity?
 }
